move the pillar fetch mothed to context previder
