+++ components/PillarList.tsx
@@ -1,45 +1,20 @@
 import { ActivityIndicator, Text, View, FlatList, StyleSheet} from "react-native";
 import { useEffect, useState } from "react";
 
-import { useAuth } from '@/context/authContext'
 import { StatusBar } from 'expo-status-bar'
-import { getDocs, query, where } from 'firebase/firestore'
-import { pillarRef } from '@/firebaseConfig'
+import { useAuth } from '@/context/authContext'
 import PillarItems from '@/components/PillarItems'
+import {usePillar} from '@/context/pillarContext'
 
 export const PillarList = () => {
     const {user} = useAuth();
-    const [Pillars, setPillars] = useState<Pillar[]>([]);
-
-    interface Pillar {
-        color: string;   
-        icon: string;       
-        id: string; 
-        subPillars: string[];
-        title: string;
-        type: string;
-        userId: string;
-        // Add other fields as needed
-    }
-
+    const {getPillars, Pillars} = usePillar();
     useEffect(()=>{
         console.log("PillarList: ", Pillars)
         if(user?.id){
             getPillars();
         }
     },[])
-
-    const getPillars = async()=>{
-        //fetch pillars
-        const q = query(pillarRef, where('userId','==',user?.id));
-        const querySnapshot = await getDocs(q);
-        let data:Pillar[] = [];
-        querySnapshot.forEach(doc => {
-            data.push({...doc.data() as Pillar});
-        })
-        console.log("fetch Pillars: ", data);
-        setPillars(data);
-    }
 
     const combinedData = [...Pillars, { icon: '➕', title: 'New Pillar' }];
